toolchain-binutils.eclass: [QA] Remove USE=multislot, #584610
The USE=multislot was used only to control build-time blocker on
previous versions of binutils. However, there is no technical reason not
to have multiple binutils versions installed at build time (or run
time). Considering that the flag does not control the installed files or
the package in any other way, it is an invalid use of USE flags.